Add Permalink creation logic to FluxBB importer
Create permalinks to redirect FluxBB-style URLs of the forms
* profile.php?id={user id}
* viewtopic.php?pid={post id}
* viewtopic.php?id={topid id}
* viewforum.php?id={forum id}

In the case of FluxBB linking a specific post within a topic page, FluxBB provides example URLs with a fragment added, so the full URL is of the form `viewtopic.php?pid=<post id>#p<post id>`, causing the browser to jump down to the specified post. When processing a redirect, the browser may re-attach this fragment with the old post id, but it won't have any effect, and the discourse logic should currectly jump down and highlight the post id specified in the Permalink object.